pyproject.toml
[project]
name = "paritu"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.124.4",
    "logfire[fastapi]>=4.16.0",
    "openai-agents>=0.6.3",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]
